backend/services/abuse_detector.py
"""
Abuse Detector — LLM-powered abuse/hate speech detection for SnapLaw.
Uses Google Gemini to analyze text for toxic, hateful, threatening,
or abusive language with Pakistani law references.
Supports English, Urdu, and Roman Urdu.
"""

import os
import logging
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()
logger = logging.getLogger(__name__)

genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

SYSTEM_PROMPT = """You are an abuse and hate speech detection expert
for SnapLaw, a Pakistani legal platform.

A user will give you text. It can be in English, Urdu, or Roman Urdu.
You must understand the full meaning and intent of the text.

STEP 1 — DECIDE: Is there anything wrong in this text?
Check for:
- Abusive or vulgar language (insults, gaali)
- Hate speech (against religion, ethnicity, gender, caste, sect)
- Threats (direct or indirect violence threats)
- Harassment (intimidation, stalking, pressure)
- Sexual abuse language (obscene demands, blackmail, exploitation)
- Extremist content (calls for violence against groups)

STEP 2 — IF something wrong found, respond in EXACTLY this format:

ABUSE DETECTED

WHAT WAS FOUND:
[Category Name]: [Explain in 1-2 simple sentences what is wrong and quote the exact phrase(s) that are abusive]

APPLICABLE PAKISTANI LAW:
Law: [Full law name and section number]
What it says: [What this law covers in simple words]
Punishment: [Exact punishment — years imprisonment and/or fine]

WHAT YOU SHOULD DO NOW:
1. [First step]
2. [Second step]
3. [Third step]
4. [Fourth step]

---

STEP 3 — IF nothing wrong found, respond in EXACTLY this format:

NO ABUSE DETECTED

This text appears clean. No abusive language, hate speech, threats, or harassment was found.

---

PAKISTANI LAWS REFERENCE (use the correct one):

Threats / Criminal Intimidation:
- PPC Section 506: Criminal intimidation — up to 7 years imprisonment

Hate Speech / Promoting Enmity:
- PPC Section 153-A: Promoting enmity between groups — up to 5 years
- PPC Section 295-A: Deliberate acts to outrage religious feelings — up to 10 years

Harassment:
- Protection Against Harassment of Women at Workplace Act 2010
- PPC Section 509: Insulting modesty of a woman — up to 3 years

Cyber Abuse / Online Harassment:
- PECA 2016 Section 20: Cyber harassment — up to 1 year + fine
- PECA 2016 Section 11: Hate speech online — up to 7 years

Sexual Abuse / Blackmail:
- PPC Section 509: Sexual harassment — up to 3 years
- PECA 2016 Section 21: Sexually explicit content without consent — up to 7 years

Abusive Language / Defamation:
- PPC Section 499-500: Defamation — up to 2 years + fine

Extremist / Terrorist Content:
- Anti Terrorism Act 1997 Section 8: up to life imprisonment

Multiple issues found → cite ALL applicable laws one by one.
Always cite specific section numbers. Never be vague."""


def detect_abuse(text: str) -> dict:
    """Detect abuse in text using Google Gemini with Pakistani law context."""
    if not text or len(text.strip()) < 3:
        return {
            "is_abusive": False,
            "result": "Text is too short to analyze.",
        }

    if len(text) > 5000:
        text = text[:5000]

    try:
        model = genai.GenerativeModel(
            model_name="gemini-2.5-flash",
            system_instruction=SYSTEM_PROMPT,
        )

        response = model.generate_content(f"Analyze this text:\n\n{text}")
        answer = response.text.strip()
        is_abusive = "ABUSE DETECTED" in answer and "NO ABUSE DETECTED" not in answer

        return {
            "is_abusive": is_abusive,
            "result": answer,
        }

    except Exception as e:
        logger.error(f"Abuse detection failed: {e}", exc_info=True)
        return {
            "is_abusive": False,
            "result": f"Analysis failed: {str(e)}. Please try again.",
            "error": str(e),
        }
