backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
transformers>=4.44.0
torch>=2.6.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
openai>=1.45.0
supabase>=2.7.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
python-dotenv>=1.0.0
numpy>=1.26.0
Pillow>=10.0.0
pytesseract>=0.3.10
PyMuPDF>=1.24.0
google-generativeai>=0.8.0
